refactor controller logic into service layer

+++ src/main/java/cat/uvic/xips/services/UserServiceImpl.java
@@ -6,6 +6,7 @@ import cat.uvic.xips.entities.User;
 import cat.uvic.xips.entities.UserProfile;
 import cat.uvic.xips.repositories.UserRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.Setter;
 import lombok.SneakyThrows;
 import okhttp3.MediaType;
 import okhttp3.OkHttpClient;
@@ -14,9 +15,7 @@ import okhttp3.RequestBody;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import java.io.IOException;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -24,11 +23,12 @@ import java.util.UUID;
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
-    private final OkHttpClient client = new OkHttpClient();
+    @Setter
+    private  OkHttpClient client = new OkHttpClient();
 
 
     @Override
-    public void save(UserCreationRequest userCreationRequest) {
+    public User save(UserCreationRequest userCreationRequest) {
 
         UserProfile userProfile = UserProfile.builder()
                 .firstname(userCreationRequest.getFirstName())
@@ -49,7 +49,12 @@ public class UserServiceImpl implements UserService {
 
         userProfile.setUser(user);
 
-        userRepository.save(user);
+        return userRepository.save(user);
+    }
+
+    @Override
+    public User save(User user) {
+        return userRepository.save(user);
     }
 
     @Override
@@ -73,11 +78,8 @@ public class UserServiceImpl implements UserService {
                 .build();
 
 
-        try (var okhttpResponse = client.newCall(request).execute()) {
-            return okhttpResponse;
-        } catch (IOException e) {
-                throw new RuntimeException();
-        }
+        return client.newCall(request).execute();
+
     }
 
     @Override
@@ -85,36 +87,45 @@ public class UserServiceImpl implements UserService {
         return userRepository.findAll();
     }
 
-    @Override
-    public User findByUsername(String username) {
-        return userRepository.findByUsername(username).orElseThrow(() -> new UsernameNotFoundException("User not found: " + username));
-    }
 
-    @Override
-    public Optional<User> findUserById(UUID id) {
-        return userRepository.findById(id);
-    }
+    public void deleteUser(String username, UUID id) {
 
-
-    @Override
-    public void deleteByUsername(String username) {
-        User user = userRepository.findByUsername(username).orElseThrow(() -> new UsernameNotFoundException("User not found: " + username));
-        userRepository.delete(user);
-    }
-
-    @Override
-    public void deleteById(UUID id) {
-        if (!userRepository.existsById(id)) {
-            throw new UsernameNotFoundException("User not found");
+        if(username == null && id == null) {
+            throw new IllegalArgumentException("should have smth in params");
         }
-        userRepository.deleteById(id);
-    }
 
+        if (username != null) {
+            User user = userRepository.findByUsername(username)
+                    .orElseThrow(() -> new UsernameNotFoundException("User not found: " + username));
+            userRepository.delete(user);
+        }
+
+        if (id != null) {
+            User user = userRepository.findById(id)
+                    .orElseThrow(() -> new UsernameNotFoundException("User not found: " + id));
+            userRepository.delete(user);
+        }
+    }
     @Override
     public void setRating(Rating rating) {
         User user = rating.getUser();
         user.getRatings().add(rating);
         userRepository.save(rating.getUser());
+    }
+
+
+    public User findUser(String username, UUID id) {
+
+        if (username == null && id == null) {
+            throw new IllegalArgumentException("error");
+        }
+
+        if (username != null) {
+            return userRepository.findByUsername(username).orElseThrow(() -> new UsernameNotFoundException("User not found: " + username));
+        }
+
+        return userRepository.findById(id).orElseThrow(() -> new UsernameNotFoundException("User not found: " + id));
+
     }
